requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==44.0.1
daphne==4.1.2
defusedxml==0.7.1
distlib==0.3.9
Django==5.1.6
django-allauth==65.4.1
django-channels==0.7.0
django-cors-headers==3.14.0
django-ninja==1.3.0
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
filelock==3.17.0
frozenlist==1.4.1
humanize==4.11.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
multidict==6.1.0
oauthlib==3.2.2
Pillow==9.5.0
platformdirs==4.3.6
propcache==0.2.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
pyOpenSSL==25.0.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2023.3
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.4.4
stripe==11.1.0
twilio==9.3.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.2.3
virtualenv==20.29.2
yarl==1.15.4
zope.interface==7.2
